requirements.txt
langchain
langchain-community
langchain-openai
langgraph
chromadb
pypdf
python-dotenv
openai